+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 
 (Dates below are YYYY/MM/DD)
 
-v0.9.5 - TBD
+v1.0.0 - TBD
 -------------------
-- **Enhancement:** Add ```--ddoc=dir``` to also generate a DDOC macro file.
+- **Change:** The generated ```packageTimestamp``` is changed from [ISOExt](http://dlang.org/phobos/std_datetime.html#toISOExtString) format to human readable. The ISOExt formatted version is now called ```packageTimestampISO```.
 - **Enhancement:** Add ```-r|--root``` to support projects in any directory, not just the current directory.
 - **Enhancement:** Minor improvements to ```--verbose``` and ```--trace``` outputs.
 - **Fixed:** Don't update the version file (and thus trigger a project rebuild) if the version file doesn't need updated. Bypass this check with the new ```--force``` flag.